+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### Added:
 - Added emi fletching recipe category
+- Added teleport scroll
+- Added teleport potion
 ### Fixed:
 - 
 ### Changed: